+++ test/integration/redirections/spend_limits.test.js
@@ -7,7 +7,12 @@ const { expect } = chai;
 describe('# Spend Limits - Integration Tests', () => {
    it('should get a redirection URL to change the spend limits', async () => {
       const appId = '1234567890';
-      const handCashConnect = new HandCashConnect(appId, Environments.iae);
+      const appSecret = process.env.app_secret;
+      const handCashConnect = new HandCashConnect({
+         appId,
+         appSecret,
+         env: Environments.iae,
+      });
       const redirectUrl = 'https://mysite.com';
       const changeSpendLimitsUrl = handCashConnect.getChangeSpendLimitsUrl(redirectUrl);
 
@@ -23,7 +28,12 @@ describe('# Spend Limits - Integration Tests', () => {
 
    it('should get a redirection URL to change the spend limits without redirection url', async () => {
       const appId = '1234567890';
-      const handCashConnect = new HandCashConnect(appId, Environments.iae);
+      const appSecret = process.env.app_secret;
+      const handCashConnect = new HandCashConnect({
+         appId,
+         appSecret,
+         env: Environments.iae,
+      });
       const changeSpendLimitsUrl = handCashConnect.getChangeSpendLimitsUrl();
 
       expect(changeSpendLimitsUrl)