+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.4.0] - 2022-03-17
+- Change SDK initialization to define `appSecret`.
+- Revert `clientUrl` so it redirects to app.handcash.io instead of handcash.io.
+
 ## [0.3.1] - 2022-02-06
 - Extend `Wallet` to expose the user total balance via `account.wallet.getTotalBalance()`.